adding documentation links to transferRepresentation, automerge type, and how to define the type in your app
